implement project search filter
